ImproperCTypes: move code and tests into proper directories
