+++ tests/ui/lint/improper-ctypes/allow-phantomdata-in-ffi.rs
