+++ tests/ui/lint/improper-ctypes/lint-113436-1.rs
