+++ tests/ui/lint/improper-ctypes/lint-73249-2.rs
