+++ tests/ui/lint/improper-ctypes/lint-73249-3.rs
