+++ tests/ui/lint/improper-ctypes/lint-73249-5.rs
