+++ tests/ui/lint/improper-ctypes/lint-73251-1.rs
