+++ tests/ui/lint/improper-ctypes/lint-73251-2.rs
