+++ tests/ui/lint/improper-ctypes/lint-ctypes.stderr
