+++ tests/ui/lint/improper-ctypes/lint-enum.rs
