+++ tests/ui/lint/improper-ctypes/lint-fn.rs
