+++ tests/ui/lint/improper-ctypes/lint-non-recursion-limit.rs
