+++ tests/ui/lint/improper-ctypes/mustpass-113436.rs
