+++ tests/ui/lint/improper-ctypes/mustpass-113900.rs
