+++ tests/ui/lint/improper-ctypes/mustpass-134060.rs
