+++ tests/ui/lint/improper-ctypes/mustpass-66202.rs
