+++ tests/ui/lint/improper-ctypes/mustpass-73249-4.rs
