+++ tests/ui/lint/improper-ctypes/mustpass-73251.rs
