+++ tests/ui/lint/improper-ctypes/mustpass-73747.rs
