+++ tests/ui/lint/improper-ctypes/repr-rust-is-undefined.stderr
